requirements.txt
finnhub-python==2.4.22
ipykernel==6.29.5
openai==1.65.5
plotly==6.0.0
python-dotenv==1.0.1
streamlit==1.42.2
yfinance==0.2.54
